+++ config/services.yaml
@@ -50,6 +50,16 @@ services:
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
 
+    App\Services\Common\MailInformer:
+        arguments:
+            $email_admin: '%env(EMAIL_ADMIN)%'
+            $email_transfer: '%env(EMAIL_TRANSFER)%'
+
+    App\Services\Common\TelegramInformer:
+        arguments:
+            $tm_token: '%env(TELEGRAM_BOT_TOKEN)%'
+            $tm_user_id: '%env(TELEGRAM_USER_ID)%'
+
     App\Twig\:
         resource: '../src/Twig'
         public: false
@@ -93,8 +103,13 @@ services:
         tags:
             - { name: doctrine.orm.entity_listener, event: preUpdate, entity: App\Entity\User }
 
-    App\EventListener\OrderListener:
+    App\EventListener\OrderProductListener:
         arguments: ['@parameter_bag']
         tags:
             - { name: doctrine.orm.entity_listener, event: postPersist, entity: App\Entity\OrderProduct }
             - { name: doctrine.orm.entity_listener, event: preRemove, entity: App\Entity\OrderProduct }
+
+    App\EventListener\OrderListener:
+        arguments: ['@translator.default']
+        tags:
+            - { name: doctrine.orm.entity_listener, event: postPersist, entity: App\Entity\Orders }
