+++ src/EventListener/OrderListener.php
@@ -3,46 +3,28 @@
 
 namespace App\EventListener;
 
-use App\Entity\OrderProduct;
+use App\Entity\Orders;
+use App\Services\UpdateManager;
 use Doctrine\ORM\Event\LifecycleEventArgs;
-use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
+use Symfony\Bundle\FrameworkBundle\Translation\Translator;
 
 class OrderListener
 {
-    private $parameterBag;
+    private $updateManager;
+    private $translator;
 
-    public function __construct(ParameterBagInterface $parameterBag)
+    public function __construct(Translator $translator, UpdateManager $updateManager)
     {
-        $this->parameterBag = $parameterBag;
+        $this->updateManager = $updateManager;
+        $this->translator = $translator;
     }
 
-    public function postPersist(OrderProduct $orderProduct, LifecycleEventArgs $args)
+    public function postPersist(Orders $order, LifecycleEventArgs $args)
     {
-        $image = $orderProduct->getImage();
+        $message = $this->translator->trans('Создан новый заказ');
+        $message .= ': ' . $order->getId() . '; ';
+        $message .= $this->translator->trans('Заказчик') . ': ' . $order->getFullName() . ' (' . $order->getCity() . ')';
 
-        if (!empty($image)) {
-            $image_item_dir = $this->parameterBag->get('image_item_dir');
-            $absolute_path_item = $image_item_dir . "/" . $image;
-            if (file_exists($absolute_path_item)) {
-                $image_order_dir = $this->parameterBag->get('image_order_dir');
-                $absolute_path_order = $image_order_dir . "/" . $image;
-
-                copy($absolute_path_item, $absolute_path_order);
-            }
-        }
-    }
-
-    public function preRemove(OrderProduct $orderProduct, LifecycleEventArgs $args)
-    {
-        $file_name = $orderProduct->getImage();
-
-        if (!empty($file_name)) {
-            $image_order_dir = $this->parameterBag->get('image_order_dir');
-            $absolute_path = $image_order_dir . "/" . $file_name;
-
-            if (file_exists($absolute_path)) {
-                unlink($absolute_path);
-            }
-        }
+        $this->updateManager->notifyOfUpdate($message, ['mail', 'telegram']);
     }
 }
