+++ translations/messages.uk.yaml
@@ -277,3 +277,5 @@ API ключ %num%: ''
 за: ''
 шт: ''
 Стоимость заказа: ''
+Создан новый заказ: 'messages'
+Заказчик: 'messages'
